made components  about head, foot, github.
