+++ app/blog/page.tsx
@@ -1,72 +1,12 @@
-
-import Image from 'next/image'
 import styles from './page.module.css'
 import Link from 'next/link';
-import { Button, Flex, Heading, Input, position, useColorMode, useColorModeValue} from "@chakra-ui/react";
+import { Flex} from "@chakra-ui/react";
+
+import Header from "@/app/src/components/Header"
+import Footer from '../src/components/Footer';
 
 
 export default function Home() {
-    const GoToGithub = () => {
-        return (
-          <Link href="https://github.com/LastWord69ANIMA"
-          >
-            <Image
-                src={"/github-icon.svg"}
-                alt={"Picture of Github"}
-                width={50}
-                height={100}
-            ></Image>
-          </Link>
-        )
-    }
-
-    const Header  = () => {
-        return(
-            <header className={styles.header}>
-            <Link
-                href="/"
-                
-            >
-                <h1>
-                    家
-                </h1>
-            </Link>
-
-            <Link
-                href="https://micro-cms-tutorial-seven.vercel.app/"
-            >
-                <h1>
-                    事
-                </h1>
-            </Link>
-
-            <Link
-                href="/profile"
-            >
-                <h1>
-                    自
-                </h1>
-            </Link>
-
-            <Link 
-                href="/contact"
-            >
-                <h1>
-                    問
-                </h1>
-            </Link>
-        </header>
-        )
-    }
-
-    const Footer = () => {
-        return(
-            <footer className={styles.footer}>
-            <GoToGithub />
-            <p>Copyright ©Lastword69ANIMA ※著作権はありませんが、一応ここに記録。</p>
-        </footer>
-        )
-    }
   return (
     <div
     className={styles.Isometric}
@@ -78,7 +18,7 @@ export default function Home() {
 
     <div>
         <Flex className={styles.prehome}>
-            現在、作業中。
+            ※次回、microCMSのAPIを使用する予定
         </Flex>
 
         <Flex className={styles.prehome}>
