+++ app/src/components/Header.tsx
@@ -1,0 +1,43 @@
+import styles from './page.module.css'
+import Link from 'next/link';
+
+const Header  = () => {
+    return(
+        <header className={styles.header}>
+        <Link
+            href="/"
+            
+        >
+            <h1>
+                家
+            </h1>
+        </Link>
+
+        <Link
+            href="https://micro-cms-tutorial-seven.vercel.app/"
+        >
+            <h1>
+                事
+            </h1>
+        </Link>
+
+        <Link
+            href="/profile"
+        >
+            <h1>
+                自
+            </h1>
+        </Link>
+
+        <Link 
+            href="/contact"
+        >
+            <h1>
+                問
+            </h1>
+        </Link>
+    </header>
+    )
+}
+
+export default Header;
